Remove :scalastyle_integration from whitelist
This fails currently because it does not have pants.ini in the build root. We need pants.ini to know to skip scalafmt and scalafix. However, adding pants.ini results in complaints that BUILD.tools is not included, which then results in complaints that contrib/scalajs/3rdparty is not included. In a followup PR we need to figure out how to properly include pants.ini.