`pants.pex` used for tests always includes `pants.ini` (#8260)
Any integration test that depended on `//:pants_ini` would fail because it would fail to resolve `BUILD.tools`. Adding `BUILD.tools` results in a failure that `contrib/scalajs/3rdparty` is not included. Adding `contrib/scalajs/3rdparty` then results in an Ivy failure to resolve, which ended up being because the Ivy settings were not being copied from `build-support` into the chroot.

Beyond this PR getting ITs to properly work when `//:pants_ini` is included, this PR intentionally makes it so that _every_ IT has `pants.ini` (and `BUILD_ROOT`) included in the PEX. This was found to be necessary while working on https://github.com/pantsbuild/pants/pull/8259 in order to get certain ITs to work properly, such as `test_scalastyle_integration.py` needing it so that the test skips scalafmt and scalafix.