Port `tasks`, `base`, and `awslambda` integration tests to V2 (#8259)
These tests simply needed to make explicit their dependencies on `testprojects` and `examples`.

### Convention for depending on `testprojects` & `examples`

This PR establishes a new convention for how to depend on `testprojects` and `examples`, which had never been done until https://github.com/pantsbuild/pants/pull/8257. Typically, the projects must be declared as `files()` to ensure that the source root is not stripped. Specifically, `rglobs()` can be used to get all files for the project, including the `BUILD` files, so that the chroot has everything needed for Pants to run against the target.

The main question is where should the `files()` declaration live and how granular should it be? On one hand, we could declare `rglobs('*')` at the root of `testprojects` and have every IT depend on this to never have to worry about a missing `testproject`; the downside is that this brute force slows down our tests because we must copy dramatically more files than necessary and it becomes more likely the cache will be invalidated unnecessarily. 

On the other end of the spectrum, we could get extremely granular in giving a `files()` entry for each file. This would obviously be too tedious and finicky, but it would ensure we never copy more files than necessary.

Instead, this establishes the convention of creating a `BUILD` file one level above the directory in question and creating an entry called `[folder_name]_directory`, e.g. `examples/src/java/org/pantsbuild/example/hello` has an entry in its parent folder's `BUILD` called `hello_directory`, which will recursively glob everything in the `hello` folder. Any IT that uses this particular testproject would then depend on `examples/src/java/org/pantsbuild/example:hello_directory`. 

### Also redistributes CI shards
Now 7% of IT targets are ran with remote execution, so we go from 20 to 19 shards for V1. We set `travis_wait 20` because many of these integration tests take a long time.