Use InputFilesContent instead of InputFileContent (#8272)
### Problem

cf. the discussion on https://github.com/pantsbuild/pants/pull/8226 , it makes more sense for the input type for the Digest-yielding intrinsic to be a wrapper around `FilesContent` rather than `FileContent`. This makes the intrinsic more general - it can be used to create a `Digest` of several files at once.

### Solution

Rename the type `InputFileContent` to `InputFilesContent` everywhere it is used, change the v2 engine intrinsic to handle multiple input `FileContent` objects, and update tests accordingly.

### Result

This commit shouldn't change any end-user-facing behavior.