Introduce multi-platform support to hermetic execution. (#8216)
### Problem

We recently added execution request speculation to allow greater task throughput in the pants v2 engine. Currently this feature is disabled by default because remoting in general is off by default. When we turn remoting on we want speculation to just work. Our current execution request model does not include the concept of a platform. Therefore v2 rules have no way of declaring their execution and target platforms, and the engine has no way of discerning which platforms it has available to it.

### Solution

This large change set adds platform awareness to the pants v2 rule graph and the underlying engine. From the top down we:
* add a new request type called `MultiPlatformExecuteProcessRequest`, which is a dictionary like datatype in python land, and a hashmap in rust land.
* Adds new @rules which allow mapping from `MultiPlatformExecuteProcessRequest`s to results, in a similar way to simple `ExecuteProcessRequests`
* Adds additional methods to the `process_execution::CommandRunner` trait which allow a node to decide if a request is compatible with the available platform(s) *before* scheduling it to run. i.e it fails fast.
* up-casts all requests in the engine into `MultiPlatformExecuteProcessRequests` so that we only have to deal with a single code path and data type which we pass around.
* Makes decisions in the `SpeculatingCommandRunner` regarding which execution strategies are available for the passed request, and speculating if possible.


### Result

Users can expect platform dependent tasks to run on the proper platform. Rule writers can add rules which emit different requests depending on the platform that are intended to execute on. Currently cross building is *not* supported even though we introduce the concept of a target platform, only the execution platform is matched when deciding which request to run where. The target platform is assumed to be the platform where the pants client is running.