Correct misleading comment in graph::Walk
The comment claimed the walk was topological, which lead me to try and
use it as I needed to walk the path topologically.
It actually isn't as we don't check that the neighbors have no other
dependency that is ahead of them in the topological order before adding
them to the queue.

I went over the callers and verified that none of them relied on a
topological ordering, which means this hasn't actually manifested into a
bug (yet).

Fix the comment to avoid clients having the wrong expectation.