Simplify critical-path algorithm
This addresses a code review comment on PR 8308.

Instead of adding a common sink for each roots, pick the longest path
from the output of the Bellman-Ford algorithm.
This means the weights of the paths must refer to the duration of the
target node instead of the duration of the source node.