Introduce `hash_utils.hash_dir`. (#8327)
Implement a simple stable hash for the recursive contents of a directory
for use in v1 Tasks. In v2 we have the `fs/store` crate and intrinsics
that expose its types to @rules; this just tides v1 Tasks over for cases
where we just need a hash and not a Snapshot we'll never materialize
(via `self.context._scheduler.<adhoc sync apis>`). Performance is
~identical to the `self.context._scheduler.capture_snapshots` API (a
few percent faster on average).

This change supports #8263.