Add Black config for pants
Black is opinionated, which means that not much config is necessary.

Provide the minimum needed for pants' own taste:

line length of 100, which is consistent with previous practice (see
pre-existing .isort.cfg for a piece of evidence backing this statement)
exclude a few directories such as vitrualenv that really shouldn't be
touched by Black. We also exclude "compilation_failure". This pattern
matches a single file in our repo which trips Black up with its mix of
invalid python syntax and use of a unicode character.

When we run Black later, through pants or manually, it should just do
what we want without needing any CLI arguments.