Make an error message more end-user friendly.
If a non-test target is used in a context expecting a test target,
we want the error message to be meaningful to the end user.

The previous message used "pants developer" terminology, referring
to the target type not being a member of the union.

The new message references the target address, and doesn't mention
the "union" concept, which in the opinion of this author is more
in line with how end users understand things.