Write integration tests for pants fmt with black
Use existing example projects as input and assume they're
already formatted because that will be an easy invariant to maintain
once we enforce black formatting of all files.

Testing the actual formatting wouldn't bring a lot of value as we can
rely on black behaving properly in general, and we don't need to couple
these tests too much to black specific things.

These tests prove that black is indeed called on the right files and
with the right arguments in a few contexts, which covers the pants
integration of black.