Introduce Workspace type to wrap directory materialize (#8329)
### Problem

The current way we materialize directories to disk involves calling a python method on the scheduler that wraps a FFI rust function (`materialize_directories`). This is awkward for v2 rules - we would like to be able to request the directory-materialization functionality if and only if we need it as the input to a rule.

### Solution

Create a new type `Workspace`, which has a handle to the rust FFI function, and can be requested as rule input. Eventually, we would like to enforce that this `Workspace` type can *only* be requested in `@console_rules`, but this is future work. A `Workspace` object has a method `materialize_directories` that can be called with a tuple of `DirectoryToMaterialize` objects, same as the current FFI. It returns an object of type`MaterializeDirectoriesResult`, which wraps several `MaterializeDirectoryResult`s, representing what files have actually been written to disk.