Add support for running Pants with Python 3.8 (#8415)
Python 3.8 will be released on October 14. This ensures that Pants will work with Python 3.8, as the current version of `cffi` could not be installed with python3.8.rc0.

Because we mark the abi for `pantsbuild.pants` as `cp36.abi3`, we do not need to release a new wheel; the current wheel will work with Python 3.8.