[junit-run] apply platform args to junit run java exec (#8421)
### Problem
When a test is using a particular jvm platform, it should also use that platform's args. See #8419

### Solution

This patch changes the junit run task to include those args in the jvm_options kwarg passed to execute the test suite.

### Result

The args are now passed to Java.

### Follow on work
I think this isn't the only place with this issue. I found it in jvm run as well.