Add `@frozen_after_init` decorator to make some @dataclass uses safer with V2 (#8431)
### Problem
In https://github.com/pantsbuild/pants/pull/8423#issue-325939000, we realized the need to be able to not always use `frozen=True` when using dataclasses for the V2 engine. Specifically, when we need a custom `__init__()`, we cannot have `frozen=True` because `__init__()` must be able to mutate the object's attributes.

However, we still want to keep the benefits of `frozen=True` as much as possible. After we create the object, we do not want to allow any additional modifications. This is important because the V2 engine uses the `__hash__()` of an object to determine its node ID - if the object mutates after creation, the ID will become invalid.

### Solution
As suggested in https://github.com/pantsbuild/pants/pull/8423#issuecomment-539810058, introduce a new decorator `@frozen_after_init` that only allows setting attributes in the `__init__()` and then marks the object as frozen after.

This is done by setting a flag in the `__init__()` to mark that the object should there-on-out be frozen. Then, we modify `__setattr__()` to check for that flag to prohibit any mutations other than those in the `__init__()`.

### Result
`@dataclass(unsafe_hash=True)` can now be marked with `@frozen_after_init` so that we can have the custom `__init__()`s but not lose the benefits of `frozen=True`.

This also works on any arbitrary class.

We do not yet enforce that dataclasses _must_ use this attribute (or `frozen=True`), which is left for https://github.com/pantsbuild/pants/pull/8181.