Allow ignoring flake errors by name in addition to code. (#8444)
### Problem

The error codes assigned to pyflakes error classes are fragile: there are 48 error classes, of which we have codes assigned to just a handful. When a message has not yet been classified, the code `F999` is used, but the effect of ignoring `F999` is that you ignore "everything", which is risky.

### Solution

Allow pyflake messages to be ignored by the error class name in addition to by their code.

### Result

Less fragility and risk as pyflakes are added.