Improve support for `Enum`s as option types (#8458)
### Problem
While we can currently use `enum.Enum` as an option and everything works automatically with default values, the experience is less than ideal when the user types an incorrect value that is not defined in the Enum. See https://github.com/pantsbuild/pants/pull/8451#discussion_r334075761.

### Solution

#### Changes to `parser.py`
Update `parser.py` to better support incorrect values for `Enum` types. This only requires:

1) Update `to_value_type()` to check for `ValueError`, which is what `Enum` throws upon an unrecognized value
2) Update `check()` to also check for subclasses of `Enum`.

Finally, add tests to `test_options.py` to ensure that we correctly support `enum.Enum` types.

#### Changes to `hash_utils.py`
We must register an Enum as a valid type. We hash against the enum's `value`, not the symbol `name`. This is an arbitrary choice. Either should provide a stable hash given the singleton property of the enum.

### Result
An invalid enum option now triggers an exception closer to what we get when this happens with `enum()`:

```
$ ./pants --glob-expansion-failure=foo

...
pants.option.errors.ParseError: Error applying type 'GlobMatchErrorBehavior' to option value 'foo', for option '--glob_expansion_failure' in global scope: 'foo' is not a valid GlobMatchErrorBehavior
```