Use `partially_type_checked` tag for incremental type hinting of Pants repo (#8464)
### Problem
Currently, to have MyPy run against a target, you must mark it with `type_checked` or it must be a dependency of a target marked with this tag.

Often, Pants devs want to add 1-2 type hints but don't have time to type check the whole target. So, they must decide if they should still add `type_checked` even though the target is not fully typed. If they add it, it now becomes much harder for us to track what work remains.

We want some way to:

1. Have MyPy check partially typed targets, even if the target is not entirely typed.
1. Track what is partially vs. fully typed.

### Solution
Introduce the `partially_type_checked` tag and have `mypy.py` run against any targets with `partially_type_checked` or `type_checked`.

### Result
The barrier to entry becomes much lower to adding some type hints to a target and enforcing that they work.

This PR allows us to pursue a new strategy for the type hint migration (https://github.com/pantsbuild/pants/issues/6742): get as many targets as possible to work with `partially_type_checked`. **Even if a target has zero type hints, we should add `partially_type_checked` to it** so that we enforce that we do not introduce any new errors and so that when we add do add type hints, we know the rest of the codebase is using things correctly.