Fix async_semaphore deadlock, and only fail speculation if primary command fails. (#8374)
### Problem

Speculation should not interfere with a users current experience of pants. Currently if either one of the dispatched commands fails the failure is propagated to the user. Additionally, we've observed a deadlock in the presence of speculation above `BoundedCommandRunner`.

### Solution

Instead of failing fast on the first returned error. If the secondary request fails before the primary request we let the primary request finish and take the result/error from that request. I.e we only use successful results from the secondary.

### Result

If the user has enabled speculation, only failures from the primary command runner, which default to the local machine, will be fatal. Fixes #8089.