Remove `enum()` to consolidate around standard library Enum (#8475)
Now that we are no longer using `enum()` in favor of `collections.Enum`, we can safely remove `enum()` and `ChoicesMixin`. This allows us to converge around one and only one way to do things.

Normally, we would use a deprecation cycle, but we decided in https://github.com/pantsbuild/pants/pull/8467 that the V2 engine API abstractions were always experimental so do not follow the normal `:public API` rules. By avoiding a deprecation cycle, we can get to MyPy working with our codebase much quicker as the only remaining major blocker is `datatype()`.

### Also adds new options test
We found that the global options did not serialize properly to JSON when using the old `enum()` type. This test works now that we use `collections.Enum`. We add the test to prevent any regressions.