Add workflow option for -Youtline as an outliner (#8427)
### Problem

Starting with `scalac` `2.12.9`, there is a `-Youtline` option to produce `.sig`
files, similar to `rsc`'s "header" `.class` files. Unlike `rsc`, `-Youtline`
supports full Scala, making it much more of a plug-and-play solution. Thus we
want to add an option for using `-Youtline` as the outliner.

### Solution

Add a new JVM compile workflow option, `outline-and-zinc`, which will act like
`rsc-and-zinc` but using `scalac -Youtline` as the outlining technology. In
addition, `wartremover` will be used to enforce the `PublicInference` rule, as
type ascriptions will have a significant performance impact on outlining.

### Result

Users will be able to use `-Youtline` in their builds.