Increase our client count and sync bootstrapped files. (#8536)
### Problem

We increased the RBE worker pool size to improve the runtime of the integration test shard, but I did not increase the client count (based on the assumption that some of the delay was due to contention between multiple travis runs, rather than due to straight line concurrency of one travis shard).

Additionally, we have been seeing periodic SIGBUS errors while starting pants immediately after bootstrapping it.

### Solution

Increase our client count to match our RBE worker pool size. If we have multiple travis runs going on at once, some of them might starve, but this at least increases the chances of saturating the backend.

Additionally, based on the hypothesis that the SIGBUS error has something to do with the bootstrap process itself not syncing files to disk, add a manual sync to leave in place until we either confirm or deny that hypothesis.