Reduce boilerplate in FFI scheduler initialization (#8522)
###Problem

There's a lot of verboseness in the way we currently initialize the scheduler over the FFI interface, and this commit is a first stab at trying to reduce some of that verboseness. Currently scheduler.py calls native.new_scheduler, which has a lengthy function signature mostly expecting Python type arguments. In turn, it passes these Python types to the Rust scheduler_create method so that the Rust engine code can create TypeIds for them.

###Solution

We can simply import the Python types we need Rust to be aware of in native.py and avoid having to pass them around in scheduler.py. This reduces the signature of native.new_scheduler to a more tractable length. It also creates a new Python struct EngineTypes that is isomorphic to the already-defined Types struct in rust, and passes the Python types to the engine this way, which reduces the amount of separate function arguments that need to be threaded around several function calls.