Add Zinc for outlining support (#8560)
### Problem

`scalac`'s built-in `-Youtline` mode can be used by `zinc` if the underlying
Scala version is at least `2.12.9`. The previous implementation of `-Youtline`
support doesn't allow this, and has an issue where `zinc` and a standalone
`scalac` causes a race condition when they are run in the same nailgun. Thus,
`scalac -Youtline` is forced to run as a subprocess, which is too slow, or in
`hermetic`.

### Solution

We introduce a `compile.rsc` flag, `--zinc-outline`.  When this flag
is true under a non-hermetic `outline-and-zinc` workflow, we follow logic
similar to `ZincCompile.compile` which will pass the necessary `-Youtline`
arguments to `zinc`.

### Result

`scalac` outlining will be able to use the same nailgun as `zinc`.