add compiler_option_sets_enabled_scalac_plugins option to JvmCompile (#8567)
### Problem

When working with @ShaneDelmore on an automated change to the twitter monorepo, we began to want to turn on and off different compiler options which belonged to an internal scalac plugin. While `compiler_option_sets_{en,dis}abled_args` are able to manually add arguments which get passed to a plugin via prefixing them with `-S-P:<plugin_name>:<arg...>`, the value of `compiler_option_sets` for some target is not able to ensure that the scalac plugin is also on the classpath.

This means that a target which wants to declare args with `compiler_option_sets` for a scalac plugin must also declare `scalac_plugins={'<plugin_name>'}`. This redundancy makes it more difficult to make automated changes, and may be mildly confusing for users.

### Solution

- Add the new `--compiler-option-sets-enabled-scalac-plugins` dict option to `JvmCompile`, which maps (a compiler option set name) -> (a list of scalac plugins that should also be registered when that option set is enabled).

### Result

A compiler option set can now implicitly require a named scalac plugin! Option sets can now be developed which span across scalac plugins as well, making them even more flexible and allowing monorepo caretakers to safely make even more changes to the build, without requiring a change in pants code.