Replace `Collection.of(T)` with typed `Collection[T]` (#8574)
Re-implements https://github.com/pantsbuild/pants/pull/8467, which had to be reverted due to issues with Python 3.7. This works with Python 3.6+.

The reason the original PR failed with Python 3.7 is that https://www.python.org/dev/peps/pep-0560/ changed the implementation of `typing.Generic` to no longer use `GenericMeta` as its metaclass. This meant that the type of something like `RegexMatchResults = Collection[RegexMatchResult]` would no longer be `GenericMeta` but `_GenericAlias`, and, critically, that `RegexMatchResults` would no longer be a `type` but instead a type alias.

This can be fixed by simply replacing this idiom

```python
RegexMatchResults = Collection[RegexMatchResult]
```

with

```python
class RegexMatchResults(Collection[RegexMatchResult]):
  pass
```

The new class is also free to define any optional properties / methods, like this:

```python
class RegexMatchResults(Collection[RegexMatchResult]):
  @property
  def failed_results:
    ...
```

### Why not use `Tuple[C, ...]`?
`RegexMatchResults = Tuple[RegexMatchResult, ...]` fails for the same reasons as `RegexMatchResults = Collection[RegexMatchResult]` does with Python 3.7. With Python 3.6, it instead fails with:

> [ERROR] panic at 'Core function `typing.Tuple[pants.engine.fs.FileContent, ...]` failed: Throw(Type Tuple cannot be instantiated; use tuple() instead, "Traceback (most recent call last):\n  File \"/Users/earellano/workspace/pants/src/python/pants/engine/native.py\", line 268, in call\n    val = func(*args)\n  File \"/opt/ee/python/3.6/lib/python3.6/typing.py\", line 1285, in __new__\n    raise TypeError(\"Type Tuple cannot be instantiated; \"\nTypeError: Type Tuple cannot be instantiated; use tuple() instead\n")', src/externs.rs:311

We could try using `class FilesContent(Tuple[FileContent, ...]): pass`, but that means we lose the API we've come to depend on for `engine.objects.Collection`, namely the constructor calling `tuple()` on the args to it. Naively changing all `Collection[C]` to `Tuple[C, ...]` results in Pants failing to run properly.