Fix V2 pytest caching issues from using FrozenSet (#8591)
### Problem
Even though frozen sets are immutable in Python, sets are never deterministic. This means that the V2 cache will almost never work properly.

Tangibly, https://github.com/pantsbuild/pants/pull/8577 broke caching for the rule `create_pex` and thus the V2 pytest runner.

### Solution
Stop using `FrozenSet` and instead use the safer `Tuple`. We still use sets in the intermediate logic because we want the de-duplication, but we then cast to a sorted tuple. The sorting is important for determinism.

### Result
Caching works again.