Fix sorting of junit tests in presence of a timeout. (#8600)
### Problem

When we timeout or when junit XML is corrupted for some reason, it can prevent us from identifying the target that failed, which causes us to default to a `None` cluster of tests.

We knew this already though, and the codepath that parsed tests accounted for it. But we had missed a case.

### Solution

Default to the empty string rather than `None` when sorting tests.