Options parsing shouldn't care that positional args are specs. (#8602)
This mostly just renames the `target_specs` fields in ArgSplitter
and Options to `positional_args`.  Consumers of parsed command
lines can know that the positional args represent target specs,
but there's no need for the cmd line parser itself to know this.

Note that in practice the arg splitter still ~knows that positional
args "look like" target specs or paths, so that it can tell them
apart from unknown goals. However we can expand the positional
arg detection heuristic in the future, to include things that
aren't paths. This change breaks out that heuristic into a method,
so that we can modify it in the future.

This change also removes the old affordance for inserting the
superfluous word `goal` in command lines. E.g.,
`./pants goal compile ...` instead of `./pants compile ...`.
That was a legacy of early versions of pants, and has not
been "a thing" for a very long time.