Allow console_rule to request a single `BuildFileAddress` (#8586)
## Problem

We want to limit the run goal to requesting a single target

## Solution

Create a new @rule that can turn Specs into a single BuildFileAddress, by checking that the BuildFileAddresses. This allows the run @console_rule to be rewritten to request only a single BuildFileAddress. The new @rule will throw an exception if multiple targets are specified on the command line for a @console_rule that requests only a single BuildFileAddress.