Port cloc to v2 (#8251)
## Problem

We didn't have a v2 version of the cloc goal.

## Solution

Add a new `fast-cloc` goal using the v2 engine.

Result

This commit doesn't yet get rid of the v1 cloc goal since we still need to remove the hardcoded URL of the upstream source of the cloc binary, which we can't easily do in the v2 engine yet. No user-facing behavior around the `cloc` goal should change yet.