Return success if V2 test runner is passed an empty target (#8630)
### Problem
Currently, passing a `python_tests` target to `./pants --no-v1 --v2 test` will result in Pytest's test-discovery triggering and trying to run over anything it discovers. This results in running tests on dependencies like `unittest2`.

Meanwhile, V1 will simply not run Pytest and return with an exit code of 0. We want the semantics to be the same.

### Solution
Early return if there are no test target source files.

We also move the source file handling logic higher in the rule's body so that this early return can trigger before the costly resolve requirements logic.

Finally, this sets up a new unit test file `test_python_test_runner`. With this now setup, it will be easier to write additional unit tests for the rule rather than adding on costly integration tests.

### Result
It's now safe to run `./pants --no-v1 --v2 test` against a target with no sources.