Add `--version` and `--pytest-plugins` options to Pytest and deprecate old options (#8619)
### Problem
The options config for the V1 `Pytest` subsystem is not ideal. Rather than a specific option for each plugin like `--unittest2-requirements`, we should have a generic `--plugins` option that is a `list` type so that users can easily add or remove plugins.

We need this to add the plugin `pytest-rerunfailures`.

### Solution
Add new options `--version` and `--pytest-plugins` to the `Pytest` subsystem. Deprecate the old options.

Set up the parsing logic as follows:

- it's invalid to specify both the new and old style
- if the old style was manually configured, entirely use the old style
- otherwise, use the new style

### Result
It's now much easier to add and remove plugins to the subsystem.

If the user uses the old style, we print:

>  [WARN] /Users/eric/DocsLocal/code/projects/pants/src/python/pants/bin/pants_loader.py:85: DeprecationWarning: DEPRECATED: option 'requirements' in scope 'pytest' will be removed in version 1.25.0.dev0.
  Use --version instead.

If the user specifies both the old and new style, we print:


> Exception message: Conflicting options for --pytest used. You provided these options in the new, preferred style: `--pytest-version`, but also provided these options in the deprecated style: `--pytest-requirements, --pytest-timeout-requirements`.
Please use only one approach (preferably the new approach of `--version` and `--pytest-plugins`).