Rename some test files from `test_foo.py` to `foo_test.py` (#8659)
This allows for us to quickly see in ordered directly listings all the tests for sibling implementations, e.g. `specs.py` and `specs_test.py`.

This was the original intent of https://github.com/pantsbuild/pants/issues/7489.

Notably, we must name ITs `foo_integration_test.py`, not `foo_test_integration.py`, for Pytest to discover the target and for our `python_tests` target type to properly glob the file. This contrasts with the proposal in #7489.